+++ tests/test_on_data/ecephys/test_sorting_interfaces.py
@@ -1,7 +1,9 @@
 from datetime import datetime
 
 import numpy as np
+from numpy.testing import assert_array_equal
 from pynwb import NWBHDF5IO
+from spikeinterface.extractors.nwbextractors import read_nwbfile
 
 from neuroconv.datainterfaces import (
     BlackrockRecordingInterface,
@@ -197,6 +199,21 @@ class TestPhySortingInterface(SortingExtractorInterfaceTestMixin):
     data_interface_cls = PhySortingInterface
     interface_kwargs = dict(folder_path=str(DATA_PATH / "phy" / "phy_example_0"))
     save_directory = OUTPUT_PATH
+
+    def check_read_nwb(self, nwbfile_path: str):
+        # Test that the max channel is correctly extracted
+        super().check_read_nwb(nwbfile_path)
+
+        # check that the max channel is correctly extracted
+        max_channel = self.interface.get_max_channel()
+        assert_array_equal(max_channel, [1, 2, 5, 5, 6, 21, 13, 13, 21, 21, 22, 22, 24])
+
+        # check that max channel was properly added to sorting extractor
+        assert_array_equal(self.interface.sorting_extractor.get_property("max_channel"), max_channel)
+
+        # check that max channels were properly added to the NWB file
+        nwbfile = read_nwbfile(file_path=nwbfile_path, backend="hdf5")
+        assert_array_equal(nwbfile.units["max_channel"].data[:], max_channel)
 
     def check_extracted_metadata(self, metadata: dict):
         assert metadata["Ecephys"]["UnitProperties"] == [